Add code so far
This is all the working code I've done so far before creating the repository. There's a bunch more broken code for connecting and authenticating with the Twitch chat WebSocket server, which I'll push here soon.

Had to move secrets out of the config class and into .NET's user secrets store, so there's a new user secrets class to handle that.